stats api models can now handle scores above int
- Fix ArrayindexOutOfBoundsexception in AbstractStatsCommand::getEndPositionThrow
- Fix stats api models can now handle scores above int

+++ src/main/java/de/timmi6790/mineplex_stats/statsapi/models/java/JavaGroupsPlayer.java
@@ -32,7 +32,7 @@ public class JavaGroupsPlayer extends ResponseModel {
     public static class Stats {
         private final String game;
         private final int position;
-        private final int score;
+        private final long score;
         private final int unix;
     }
 }
